Testing Por Capa

+++ src/main/java/co/com/ingeneo/api/controller/response/OrdenModel.java
@@ -27,7 +27,7 @@ import lombok.experimental.FieldNameConstants;
 @Relation(collectionRelation = "data")
 public class OrdenModel extends RepresentationModel<OrdenModel> {
 	
-	private Integer ordenId;
+	private Long ordenId;
 
 	private ClienteModel cliente;
